+++ src/openapi/streaming/connection/transportTypes.ts
